pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scitex-audio"
version = "0.1.0"
description = "Text-to-Speech with multiple backends for scientific workflows"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.10"
authors = [
    { name = "Yusuke Watanabe", email = "ywatanabe@scitex.ai" },
]
keywords = [
    "tts",
    "text-to-speech",
    "audio",
    "speech",
    "scitex",
    "science",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "click>=8.0",
    "pydub",
    "soundfile",
    "scitex_dev",
]

[project.optional-dependencies]
pyttsx3 = ["pyttsx3"]
gtts = ["gTTS"]
elevenlabs = ["elevenlabs"]
mcp = ["fastmcp>=2.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
    "sphinx-copybutton>=0.5",
    "sphinx-autodoc-typehints>=1.25",
]
all = [
    "scitex-audio[pyttsx3,gtts,elevenlabs,mcp,dev,docs]",
]

[project.scripts]
scitex-audio = "scitex_audio._cli:main"

[project.entry-points."scitex_dev.docs"]
scitex-audio = "scitex_audio"

[project.urls]
Homepage = "https://github.com/ywatanabe1989/scitex-audio"
Repository = "https://github.com/ywatanabe1989/scitex-audio"
Documentation = "https://scitex-audio.readthedocs.io"

[tool.hatch.build.targets.wheel]
packages = ["src/scitex_audio"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "--cov=scitex_audio --cov-report=term-missing --cov-report=html:htmlcov"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
